wifi: mwifiex: handle VDLL
Some adapters send events to the Kernel requesting VDLL images. When we
receive such an event we have to upload parts of the firmware image to
the device.

When the firmware requests a VDLL image then we have to bypass the
pending command queue and have to send it immediately. Failing to do
so will cause every other command sent from the queue to timeout.

There is this check currently in mwifiex_download_vdll_block():

	if (adapter->cmd_sent) {
		mwifiex_dbg(adapter, MSG, "%s: adapter is busy\n", __func__);
		return -EBUSY;
	}

The downstream driver has this check as well. However, the downstream
driver tries to defer sending the VDLL command when this happens. It
sets a variable in the private driver struct where the main process
picks it up later.  I am not sure if this can work, because when
adapter->cmd_sent is true then it's already too late, we already missed
to send the VDLL image before the next command. In my tests I never saw
the above trigger, but this surely a place to have a closer look at.

Upstream-Status: Pending

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>