media: imx283: Provide optical black mode
The IMX283 is capable of delivering optical black regions as part of
the image capture. These regions support capture of the black levels
for calibration and are added as extra pixels on top of the full
resolution capture.

Supply an extra mode which accounts for this increased size that will
produce black regions in the output image.

Upstream-Status: Pending

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>