wifi: mwifiex: drop driver internal AP/STA limit counting
The mwifiex driver maintains internal counters to check if there
are still enough resources to change a virtual interface to a certain
type. This seems to be a remnant of old times and can be removed.

We can currently create three virtual interfaces (could be expanded to
16) and each of the interfaces can be configured to any type without
further restrictions. The limits we actually have are already correctly
described in wiphy->iface_combinations.

Upstream-Status: Pending

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>