wifi: mwifiex: fix multiple station handling
mwifiex_is_send_cmd_allowed() and mwifiex_is_tdls_chan_switching() are
called with the first bss in station mode. There can be multiple bss in
station mode and we have to consider all of them, not only the first
one. Instead of the bss priv pass the adapter to these functions and
iterate over the bss as necessary.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>