wifi: mwifiex: add function to send command specific to the adapter
Current mwifiex_send_command() takes a struct mwifiex_private * as
context pointer. There are several commands though that are specific
to the whole adapter and not to priv *. For these commands introduce
a mwifiex_adapter_send_command() function that takes the adapter as
context pointer. Some users are updated to use this function, more
will be converted in followup patches.

Upstream-Status: Pending

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>