drm/panel: add Hexagon KDU panel support
Add support for the Hexagon Keyboard-Disaplay-Unit (KDU) panel. A KDU
is a HMI addon which is equipped with:

 - a keyboard (done via MCU)
 - a touchscreen and
 - a display.

There can be two KDUs attached to the device. In such case the LVDS
signals are dupplicated in HW via a LVDS splitter. Therefore we have two
physical LVDS panel but logically there is only one. The LVDS splitter
en-/disables the output once the KDU power-supply is enabled:
 - both power-supplies present -> KDU1 and KDU2 LVDS signals enabled
 - only one power-supply present -> KDU1 or KDU2 LVDS signals enabled
 - no power-supply present -> no KDU LDVS output enabled

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>