wifi: mwifiex: drop asynchronous init waiting code
Historically all commands sent to the mwifiex driver have been
asynchronous. For this reason there is code that waits for the
last initialization command to complete before going on. Nowadays the
commands can be sent synchronously, meaning that they are completed
when the command call returns. This makes all the waiting code
unnecessary. It is removed in this patch.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>