wifi: mwifiex: pass new MAC address to mwifiex_send_cmd()
Right now setting a new MAC address in the mwifiex driver is done as:

1) mwifiex_set_mac_address() sets priv->curr_addr to the desired new MAC
   address
2) mwifiex_cmd_802_11_mac_address() (called from mwifiex_send_cmd())
   constructs the HostCmd_CMD_802_11_MAC_ADDRESS command, using the MAC
   address in priv->curr_addr
3) mwifiex_ret_802_11_mac_address(), called from the response handler,
   sets priv->curr_addr to the MAC address received with the command
   response, which of course is the same as we initially copied there
   in step 1), which makes 3) unnecessary
4) in case the HostCmd_CMD_802_11_MAC_ADDRESS command fails we have to
   revert priv->curr_addr to the old address

Simplify this to passing the new MAC address in the data_buf argument
in mwifiex_send_cmd(). The response handler will still copy the new MAC
address received with the command response to priv->curr_addr, but we no
longer have to copy the mew MAC address over to priv->curr_addr and also
do not have to revert to the old address when the command fails.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>