media: i2c: imx283: Recalculate SHR on blanking changes
The exposure control on the imx283 is handled through the SHR register.
This value is calculated based upon the hmax and vmax registers as a
property of the total line and frame length.

Ensure that the SHR is updated whenever the blankings update and adjust
the frame intervals to ensure the correct exposure is configured on the
sensor.

Upstream-Status: Pending

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>