wireless: mwifiex: increase wake timeout timer
When idling for longer while being connected to an AP it occasionally
happens that the wakeup timeout timer triggers. In this case the driver
reset the hardware and it won't come up again. Ignoring the timeout
helps, so it seems the device sometimes just takes longer to wakeup.
Increase the timeout from 3s to 10s.

Upstream-Status: Pending

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>