media: rkisp1: Set format defaults based on requested color space
When color space JPEG is requested, the ISP sets the quantization
incorrectly to limited range. To fix that, set the xfer_func, ycbcr_enc
and quantization to the defaults for the requested color space if they
are not specified explicitly. Do this only in case we are converting
from RAW to YUV.

Upstream-Status: Pending

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>