dmaengine: imx-sdma: make use of devm_add_action_or_reset to unregiser the dma_device
Make use of the devm_add_action_or_reset() to register a devm_ release
hook since the dmaengine API doesn't provide a devm_ wrapper yet.

This removes the last goto error handling within the probe function and
further trims the remove() function. Instead of freeing the irq, we can
disable it and let the devm-irq do the job to free the irq, since the
only purpose was to have the irqs disabled before calling
dma_async_device_unregister().

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>