dmaengine: imx-sdma: drop remove callback
The whole driver was converted to the devm APIs except for this last for
loop. This loop is buggy due to three reasons:
 1) It removes the channels without removing the users first. This can
    lead to very bad situations.
 2) The loop starts at 0 and which is channel0 which is a special
    control channel not registered via vchan_init(). Therefore the
    remove() always Oops because of NULL pointer exception.
 3) sdma_free_chan_resources() disable the clks unconditional without
    checking if the clks are enabled. This is done for all
    MAX_DMA_CHANNELS which hang the system if there is at least one unused
    channel.

To address the first case, we need to add fw_devlink.strict=1 to the
kernel cmdline. This enables the dma-supplier fw_devlinks and removes
the users of the sdma first before removing the sdma itself. During the
step the channel are dropped and the dmaengine core handles the
channel_resources free accordingly.

The third case is fixed with the fw_devlink.strict=1 mechansim too.

The 2nd case is fixed by not doing anything on channel0.

To sum-up, all cases are fixed by dropping the .remove() callback and
let the frameworks do their jobs.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>