dmaengine: add device_link support
Shift the device dependency handling to the driver core by adding
support for devlinks.

The link between the consumer device and the dmaengine device is
established by the consumer via the dma_request_chan() automatically if
the dmaengine driver supports it (create_devlink flag set).

By adding the devlink support it is ensured that the supplier can't be
removed while the consumer still uses the dmaengine. Furthermore it
ensures that the supplier driver is present and actual bound before the
consumer is uses the supplier.

Additional PM and runtime-PM dependency handling can be added easily too
by setting the required flags (not implemented by this commit).

The new create_devlink flag controlls the devlink creation to not cause
any regressions with existing dmaengine drivers. This flag can be
removed once all drivers are successfully tested to support devlinks.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>