net: stmmac: restart pps after set/adjust_time
LinuxPtp as well as Chrony use ptp adjust_time, set_time and adjust_freq to
adjust a phc clock. Since the stmmac doesn't adjust a pps output when
adjust_time and set_time are called, the pps out gets distorted (wrong phase,
correct fequency).

Whenever adjust_time or set_time is called, the pps output stops for 100ms and
restarts with the correct phase and frequency. This enables the usage of the
pps output as pps source for another phc.

Upstream-Status: Inappropriate [We use period out as pps out]

Signed-off-by: Bernd Sprenger <bernhard.sprenger@leica-geosystems.com>