DNI: media: dw100: Unconditionally allow dewarp map update
This patch is meant for customers having used the dynamic dewarp map
implementation without requests.

It mimics the old behavior to be backwards compatible with the old
libcamera version. Note that no measures were taken to prevent races. As
libcamera is the only user this shouldn't be an issue.

Upstream-Status: Pending

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>